refactor: 회원가입 view 부분에서 사용되는 js 코드 정리
